+++ cookbooks/cloud/recipes/owncloud.rb
@@ -3,7 +3,7 @@
 # Cookbook Name:: cloud
 # Recipe:: owncloud
 #
-# Copyright 2013,2014 Carles Muiños
+# Copyright 2013-2015 Carles Muiños
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -18,11 +18,9 @@
 # limitations under the License.
 #
 
-
 cloud = node[:apps][:cloud]
 
 # ownCloud client
-install_app "owncloud" do
+install_app 'owncloud' do
   profile cloud['profiles']['owncloud']
 end
-